Externalise utility method
